Fixed Issue with WeepingAngel Arm Poses
Fine. I'll Fix it myself. And I DID!
Don't touch them!

Minecraft uses degrees, so we convert those to radians, WE KEEP THE NEGATIVE VALUES
we also keep the floating point data because the compiler will take care of it for us, and if we round them it will make the angels off by a lot more then we think